// cat-launcher/src/generated-types/GameRelease.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GameReleaseStatus } from "./GameReleaseStatus";
import type { GameVariant } from "./GameVariant";
import type { ReleaseType } from "./ReleaseType";

/**
 * Represents a specific release of the game.
 */
export type GameRelease = { 
/**
 * The game variant this release belongs to.
 */
variant: GameVariant, 
/**
 * The version string (usually the Git tag).
 */
version: string, 
/**
 * The release notes or description.
 */
body: string | null, 
/**
 * The type of release (Stable, Experimental, etc.).
 */
release_type: ReleaseType, 
/**
 * The current installation status on the local system.
 */
status: GameReleaseStatus, 
/**
 * The date and time when the release was created.
 */
created_at: string, };